bug#81226 【运营平台】药企同步设置，勾选删除同步，未同步已删除的药品

+++ ehealth-recipe-core/src/main/java/recipe/drugTool/service/SaleDrugToolService.java
@@ -738,14 +738,18 @@ public class SaleDrugToolService implements ISaleDrugToolService {
                           }
                       }
                     }
-                Map<String, OrganDrugList> drugMap = details.stream().collect(Collectors.toMap(OrganDrugList::getOrganDrugCode, a -> a, (k1, k2) -> k1));
-                List<SaleDrugList> saleDrugListsByOrganId = saleDrugListDAO.findSaleDrugListsByOrganId(drugsEnterpriseId);
-                if (!ObjectUtils.isEmpty(saleDrugListsByOrganId)){
-                    for (SaleDrugList saleDrugList : saleDrugListsByOrganId) {
-                        OrganDrugList organDrug = drugMap.get(saleDrugList.getOrganDrugCode());
-                        if (ObjectUtils.isEmpty(organDrug)) {
-                            saleDrugListDAO.remove(saleDrugList.getOrganDrugId());
-                            deleteNum++;
+                String[] strings = config.getEnable_drug_syncType().split(",");
+                List<String> syncTypeList = new ArrayList<String>(Arrays.asList(strings));
+                if (syncTypeList.indexOf("3")!=-1){
+                    Map<String, OrganDrugList> drugMap = details.stream().collect(Collectors.toMap(OrganDrugList::getOrganDrugCode, a -> a, (k1, k2) -> k1));
+                    List<SaleDrugList> saleDrugListsByOrganId = saleDrugListDAO.findSaleDrugListsByOrganId(drugsEnterpriseId);
+                    if (!ObjectUtils.isEmpty(saleDrugListsByOrganId)){
+                        for (SaleDrugList saleDrugList : saleDrugListsByOrganId) {
+                            OrganDrugList organDrug = drugMap.get(saleDrugList.getOrganDrugCode());
+                            if (ObjectUtils.isEmpty(organDrug)) {
+                                saleDrugListDAO.remove(saleDrugList.getOrganDrugId());
+                                deleteNum++;
+                            }
                         }
                     }
                 }
@@ -892,14 +896,18 @@ public class SaleDrugToolService implements ISaleDrugToolService {
                 }
             }
         }
-        Map<String, OrganDrugList> drugMap = details.stream().collect(Collectors.toMap(OrganDrugList::getOrganDrugCode, a -> a, (k1, k2) -> k1));
-        List<SaleDrugList> saleDrugListsByOrganId = saleDrugListDAO.findSaleDrugListsByOrganId(drugsEnterpriseId);
-        if (!ObjectUtils.isEmpty(saleDrugListsByOrganId)){
-            for (SaleDrugList saleDrugList : saleDrugListsByOrganId) {
-                OrganDrugList organDrug = drugMap.get(saleDrugList.getOrganDrugCode());
-                if (ObjectUtils.isEmpty(organDrug)) {
-                    saleDrugListDAO.remove(saleDrugList.getOrganDrugId());
-                    deleteNum++;
+        String[] strings = config.getEnable_drug_syncType().split(",");
+        List<String> syncTypeList = new ArrayList<String>(Arrays.asList(strings));
+        if (syncTypeList.indexOf("3")!=-1){
+            Map<String, OrganDrugList> drugMap = details.stream().collect(Collectors.toMap(OrganDrugList::getOrganDrugCode, a -> a, (k1, k2) -> k1));
+            List<SaleDrugList> saleDrugListsByOrganId = saleDrugListDAO.findSaleDrugListsByOrganId(drugsEnterpriseId);
+            if (!ObjectUtils.isEmpty(saleDrugListsByOrganId)){
+                for (SaleDrugList saleDrugList : saleDrugListsByOrganId) {
+                    OrganDrugList organDrug = drugMap.get(saleDrugList.getOrganDrugCode());
+                    if (ObjectUtils.isEmpty(organDrug)) {
+                        saleDrugListDAO.remove(saleDrugList.getOrganDrugId());
+                        deleteNum++;
+                    }
                 }
             }
         }
